3단계 미션 제출!
refactor: ProductScreenTest product/ 패키지 이동

feat: ProductRepository 반환괎 Flow로 마이그레이션

feat: CartProduct, CartRepository

feat: Counter Component

refactor: ProductImage 컴포넌트 분리

refactor: ProductItem 컴포넌트 분리

feat: CartProductItem 컴포넌트

feat: 장바구니 화면 구현

feat: 상품을 장바구니에 담는 기능을 구현한다.

refactor: presentation 패키지에 이동

feat: 주문 완료 시 장바구니가 비워진다.

test: CartScreenTest

test: FakeRepository 적용
