+++ README.md
@@ -15,4 +15,13 @@
 - [x]: 상품 목록에서 장바구니 아이콘을 누르면 장바구니 화면으로 이동한다.  
 - [x]: 상품 상세에서 장바구니 담기 버튼을 누르면 장바구니 화면으로 이동한다.  
 - [x]: 뒤로 가기 버튼이나 아이콘을 누르면 직전 화면으로 돌아간다.  
-- [x]: 장바구니에 실제로 상품이 담기는 기능은 이 단계에서 고려하지 않는다.  
+- [x]: 장바구니에 실제로 상품이 담기는 기능은 이 단계에서 고려하지 않는다.
+
+# 3 단계 목록
+
+- [x]: 상품을 장바구니에 담는 기능을 구현한다.
+- [x]: 장바구니 화면을 구현한다.
+- [x]: 담긴 상품의 수량을 조절할 수 있어야 한다.
+- [x]: 수량을 1보다 작게 하면 장바구니에서 상품이 제거된다
+- [x]: 담긴 상품 가격의 총합이 주문하기 버튼에 표시된다.
+- [x]: 주문 완료 시 장바구니가 비워진다.
