--- app/src/main/java/nextstep/shoppingcart/MainActivity.kt
@@ -1,17 +1,0 @@
-package nextstep.shoppingcart
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            ShoppingCartTheme {
-                ShoppingNavHost()
-            }
-        }
-    }
-}
