+++ app/src/main/java/nextstep/shoppingcart/domain/model/CartProduct.kt
@@ -1,0 +1,9 @@
+package nextstep.shoppingcart.domain.model
+
+data class CartProduct(
+    val product: Product,
+    val count: Int
+) {
+    val totalPrice: Int
+        get() = product.price * count
+}
