+++ app/src/main/java/nextstep/shoppingcart/domain/model/Product.kt
@@ -1,0 +1,8 @@
+package nextstep.shoppingcart.domain.model
+
+data class Product(
+    val id: Long,
+    val imageUrl: String,
+    val name: String,
+    val price: Int
+)
