+++ app/src/main/java/nextstep/shoppingcart/presentation/MainActivity.kt
@@ -1,0 +1,17 @@
+package nextstep.shoppingcart.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import nextstep.shoppingcart.presentation.ui.theme.ShoppingCartTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            ShoppingCartTheme {
+                ShoppingNavHost()
+            }
+        }
+    }
+}
