+++ app/src/main/java/nextstep/shoppingcart/presentation/ui/component/ProductImage.kt
@@ -1,0 +1,18 @@
+package nextstep.shoppingcart.presentation.ui.component
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import coil.compose.AsyncImage
+
+@Composable
+fun ProductImage(
+    imageUrl: String,
+    modifier: Modifier = Modifier,
+    contentDescription: String? = null
+) {
+    AsyncImage(
+        model = imageUrl,
+        contentDescription = contentDescription,
+        modifier = modifier
+    )
+}
