--- app/src/main/java/nextstep/shoppingcart/product/model/Product.kt
@@ -1,8 +1,0 @@
-package nextstep.shoppingcart.product.model
-
-data class Product(
-    val id: Long,
-    val imageUrl: String,
-    val name: String,
-    val price: Int
-)
